+++ src/sys/uio.rs
@@ -2,7 +2,7 @@
 
 use crate::errno::Errno;
 use crate::Result;
-use libc::{self, c_int, c_void, off_t, size_t};
+use libc::{self, c_int, off_t, size_t};
 use std::io::{IoSlice, IoSliceMut};
 use std::os::unix::io::{AsFd, AsRawFd};
 
@@ -18,7 +18,7 @@ pub fn writev<Fd: AsFd>(fd: Fd, iov: &[IoSlice<'_>]) -> Result<usize> {
     //
     // Because it is ABI compatible, a pointer cast here is valid
     let res = unsafe {
-        libc::writev(fd.as_fd().as_raw_fd(), iov.as_ptr() as *const libc::iovec, iov.len() as c_int)
+        libc::writev(fd.as_fd().as_raw_fd(), iov.as_ptr().cast(), iov.len() as c_int)
     };
 
     Errno::result(res).map(|r| r as usize)
@@ -33,7 +33,7 @@ pub fn writev<Fd: AsFd>(fd: Fd, iov: &[IoSlice<'_>]) -> Result<usize> {
 pub fn readv<Fd: AsFd>(fd: Fd, iov: &mut [IoSliceMut<'_>]) -> Result<usize> {
     // SAFETY: same as in writev(), IoSliceMut is ABI-compatible with iovec
     let res = unsafe {
-        libc::readv(fd.as_fd().as_raw_fd(), iov.as_ptr() as *const libc::iovec, iov.len() as c_int)
+        libc::readv(fd.as_fd().as_raw_fd(), iov.as_ptr().cast(), iov.len() as c_int)
     };
 
     Errno::result(res).map(|r| r as usize)
@@ -55,7 +55,7 @@ pub fn pwritev<Fd: AsFd>(fd: Fd, iov: &[IoSlice<'_>], offset: off_t) -> Result<u
     let res = unsafe {
         libc::pwritev(
             fd.as_fd().as_raw_fd(),
-            iov.as_ptr() as *const libc::iovec,
+            iov.as_ptr().cast(),
             iov.len() as c_int,
             offset,
         )
@@ -88,7 +88,7 @@ pub fn preadv<Fd: AsFd>(
     let res = unsafe {
         libc::preadv(
             fd.as_fd().as_raw_fd(),
-            iov.as_ptr() as *const libc::iovec,
+            iov.as_ptr().cast(),
             iov.len() as c_int,
             offset,
         )
@@ -105,7 +105,7 @@ pub fn pwrite<Fd: AsFd>(fd: Fd, buf: &[u8], offset: off_t) -> Result<usize> {
     let res = unsafe {
         libc::pwrite(
             fd.as_fd().as_raw_fd(),
-            buf.as_ptr() as *const c_void,
+            buf.as_ptr().cast(),
             buf.len() as size_t,
             offset,
         )
@@ -122,7 +122,7 @@ pub fn pread<Fd: AsFd>(fd: Fd, buf: &mut [u8], offset: off_t) -> Result<usize> {
     let res = unsafe {
         libc::pread(
             fd.as_fd().as_raw_fd(),
-            buf.as_mut_ptr() as *mut c_void,
+            buf.as_mut_ptr().cast(),
             buf.len() as size_t,
             offset,
         )
@@ -181,8 +181,8 @@ pub fn process_vm_writev(
 {
     let res = unsafe {
         libc::process_vm_writev(pid.into(),
-                                local_iov.as_ptr() as *const libc::iovec, local_iov.len() as libc::c_ulong,
-                                remote_iov.as_ptr() as *const libc::iovec, remote_iov.len() as libc::c_ulong, 0)
+                                local_iov.as_ptr().cast(), local_iov.len() as libc::c_ulong,
+                                remote_iov.as_ptr().cast(), remote_iov.len() as libc::c_ulong, 0)
     };
 
     Errno::result(res).map(|r| r as usize)
@@ -216,8 +216,8 @@ pub fn process_vm_readv(
 {
     let res = unsafe {
         libc::process_vm_readv(pid.into(),
-                               local_iov.as_ptr() as *const libc::iovec, local_iov.len() as libc::c_ulong,
-                               remote_iov.as_ptr() as *const libc::iovec, remote_iov.len() as libc::c_ulong, 0)
+                               local_iov.as_ptr().cast(), local_iov.len() as libc::c_ulong,
+                               remote_iov.as_ptr().cast(), remote_iov.len() as libc::c_ulong, 0)
     };
 
     Errno::result(res).map(|r| r as usize)